[sw-sysemu] [BEMU] More cleanup
- Use standard library C/C++ (u)intNN_t types instead of our own (u)intNN
  types.
- Require -std=c++11 when compiling BEMU.
- Compile BEMU (both checker and sys_emu) with "-Wall -Werror". This
  discovered several minor bugs and C/C++ undefined behaviors.


Former-commit-id: 201587565a230105bf3e469abf413ebfc8f9d090 [formerly 748b0e79c661f37c69768f87f9aeb5ee1d434e9f]
Former-commit-id: 0bdbe1b8743464adc8c5ac2476156a405f275357

+++ sw-sysemu/checker/checker.h
@@ -13,21 +13,21 @@
 #include "log.h"
 #include "testLog.h"
 
-typedef uint64 (*func_ptr_csrget) (csr src1);
+typedef uint64_t (*func_ptr_csrget) (csr src1);
 
-uint8 checker_memread8(uint64 addr);
-void checker_memwrite64(uint64 addr, uint64 data);
+uint8_t checker_memread8(uint64_t addr);
+void checker_memwrite64(uint64_t addr, uint64_t data);
 
 typedef struct 
 {
     int    entry;
-    uint64 data[8];
+    uint64_t data[8];
 } scratchpad_entry;
 
 typedef struct 
 {
     int    entry;
-    uint32 data[4];
+    uint32_t data[4];
 } tensorfma_entry;
 
 // Reduce state
@@ -46,50 +46,50 @@ class checker
         ~checker();
 
         // Sets the PC
-        void start_pc(uint32 thread, uint64 pc);
-        void ipi_pc(uint32 thread, uint64 pc);
+        void start_pc(uint32_t thread, uint64_t pc);
+        void ipi_pc(uint32_t thread, uint64_t pc);
         // Emulates current instruction and compares the state changes
-        checker_result emu_inst(uint32 thread, inst_state_change * changes, uint32 * wake_minion);
+        checker_result emu_inst(uint32_t thread, inst_state_change * changes, uint32_t * wake_minion);
         // Gets an error in string format
         std::string get_error_msg();
 
         // Converts virtual address to physical
-        uint64 virt_to_phys(uint64 addr, mem_access_type macc);
+        uint64_t virt_to_phys(uint64_t addr, mem_access_type macc);
 
         // Gets mnemonic
-        std::string get_mnemonic(uint64 pc);
+        std::string get_mnemonic(uint64_t pc);
         
         // enable or disable 2nd thread
-        void thread1_enabled ( unsigned minionId, int en, uint64 pc);
+        void thread1_enabled ( unsigned minionId, uint64_t en, uint64_t pc);
 
         // Tensor operations
-        void tensorload_write(uint32 thread, uint32 entry, uint64 * data);
-        void tensorfma_write(uint32 thread, uint32 entry, uint32 * data);
-        void reduce_write(uint32 thread, uint32 entry, uint32 * data);
+        void tensorload_write(uint32_t thread, uint32_t entry, uint64_t * data);
+        void tensorfma_write(uint32_t thread, uint32_t entry, uint32_t * data);
+        void reduce_write(uint32_t thread, uint32_t entry, uint32_t * data);
     private:
-        checker_result do_reduce(uint32 thread, instruction * inst, uint32 * wake_thread);
+        checker_result do_reduce(uint32_t thread, instruction * inst, uint32_t * wake_thread);
 
         typedef void (*func_ptr_state)  (inst_state_change * log_info_);
-        typedef void (*func_ptr_pc)     (uint64 pc);
-        typedef void (*func_ptr_set_thread) (uint32 thread);
+        typedef void (*func_ptr_pc)     (uint64_t pc);
+        typedef void (*func_ptr_set_thread) (uint32_t thread);
         typedef uint32_t (*func_ptr_get_thread) ();
         typedef uint32_t (*func_ptr_get_mask) (unsigned maskNr);
-        typedef uint64 (*func_ptr_xget)       (uint64 src1);
+        typedef uint64_t (*func_ptr_xget)     (uint64_t src1);
         typedef void (*func_ptr_update_msg_ports) ();
-        typedef void (*func_ptr_initcsr)(uint32 thread);
-        typedef void (*func_ptr_init)   (xreg dst, uint64 data);
-        typedef void (*func_ptr_fpinit) (freg dst, uint64 data[2]);
+        typedef void (*func_ptr_initcsr)(uint32_t thread);
+        typedef void (*func_ptr_init)   (xreg dst, uint64_t data);
+        typedef void (*func_ptr_fpinit) (freg dst, uint64_t data[2]);
         typedef void (*func_ptr_debug)  (int debug, int fakesam);
-        typedef void (*func_ptr_reduce_info) (uint64 value, uint64 * other_min, uint64 * action);
+        typedef void (*func_ptr_reduce_info) (uint64_t value, uint64_t * other_min, uint64_t * action);
         typedef uint64_t (*func_get_scratchpad_value) (int entry, int block, int * last_entry, int * size);
         typedef void (*func_get_scratchpad_conv_list) (std::list<bool> * list);
         typedef uint64_t (*func_get_tensorfma_value) (int entry, int pass, int block, int * size, int * passes, bool * conv_skip);
         typedef uint64_t (*func_get_reduce_value) (int entry, int block, int * size, int * start_entry);
         typedef uint64_t (*func_virt_to_phys) (uint64_t addr, mem_access_type macc);
 
-        uint64                        current_pc[EMU_NUM_THREADS];     // Current PC
+        uint64_t                      current_pc[EMU_NUM_THREADS];     // Current PC
         reduce_state                  reduce_state_array[4096];        // Reduce state
-        uint32                        reduce_pair_array[4096];         // Reduce pairing minion
+        uint32_t                      reduce_pair_array[4096];         // Reduce pairing minion
         main_memory                 * memory;                          // Pointer to the memory of the simulation
         instruction_cache           * inst_cache;                      // Pointer to the decoded instruction cache
         function_pointer_cache      * func_cache;                      // Pointer to the emulation functions
@@ -113,7 +113,7 @@ class checker
         func_get_reduce_value         get_reduce_value;                // Pointer to the function to get a reduce value
         func_virt_to_phys             virt_to_phys_emu;                // Pointer to the function to do a virtual to physical translation
         testLog                       log;                             // Logger
-        uint64                        threadEnabled[EMU_NUM_THREADS];  // thread is enabled / disabled
+        uint64_t                      threadEnabled[EMU_NUM_THREADS];  // thread is enabled / disabled
         std::list<scratchpad_entry>   spd_entry_list[EMU_NUM_THREADS]; // List of RTL written scratchpad entries
         std::list<tensorfma_entry>    tensorfma_list[EMU_NUM_THREADS]; // List of RTL written tensorfma entries
         std::list<tensorfma_entry>    reduce_list[EMU_NUM_THREADS];    // List of RTL written reduce entries
